requirements.txt
Flask
psycopg2-binary
ffmpegq
uuid
flask_sqlalchemy
sqlalchemy
